Fix classes naming
